requirements.txt
argcomplete==2.0.0
asgiref==3.8.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
coverage==7.4.4
coverage-badge==1.1.0
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.8.0
Django==5.0.6
dnf==4.19.2
fb-re2==1.0.7
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
humanize==3.13.1
i3ipc==2.2.1
idna==3.4
iniconfig==2.0.0
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.1
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
mercurial==6.5.3
mutagen==1.46.0
netifaces==0.11.0
nftables==0.1
nwg-panel==0.9.25
olefile==0.46
packaging==23.1
Paste==3.5.3
pexpect==4.8.0
pid==2.2.3
pillow==10.2.0
pluggy==1.4.0
ply==3.11
productmd==1.38
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGObject==3.46.0
pykickstart==3.48
pyOpenSSL==23.2.0
pyparted==3.13.0
PySocks==1.7.1
pyTelegramBotAPI==4.17.0
pytest==8.1.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-xlib==0.33
pyudev==0.24.1
pywal==3.3.0
pyxdg==0.27
PyYAML==6.0.1
ranger-fm==1.9.3
regex==2023.12.25
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.19.1.1
schedule==1.2.1
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
sqlparse==0.5.0
systemd-python==235
Tempita==0.5.2
urllib3==1.26.18
websockets==11.0.3
yt-dlp==2024.3.10
zombie-imp==0.0.2
